refactor to single util file
